# internal/api/thea.openapi.yaml
openapi: 3.0.0
info:
  title: Thea Spec
  description: "Thea REST API spec. See http://github.com/hbomb79/Thea for more information"
  version: 1.0.0
  contact:
    name: Thea Support
    url: https://github.com/hbomb79/Thea
...
security:
  - permissionAuth: [] # Default security - requires authentication but no specific permissions
paths:
...
  /media:
    get:
      summary: List Media
      description: Allows a client to fetch a list of movies/series using various filtering, ordering and paging paramaters
      operationId: listMedia
      tags:
        - Media
      security:
        - permissionAuth: [media:access]
      parameters:
        - in: query
          name: genre
          description: Optional set of genres which all returned media will be associated with
          schema:
            type: array
            items:
              type: string
        - in: query
          name: allowedType
          description: Optional set of media types which can be returned by this endpoint
          schema:
            type: array
            items:
              type: string
        - in: query
          name: orderBy
          description: Optional ordering for the results, defaults to updated_at in ascending order
          schema:
            type: array
            items:
              type: string
        - in: query
          name: titleFilter
          description: Optional fuzzy title filter which all returned results must match against
          schema:
            type: string
        - in: query
          name: offset
          description: The number of items to skip before starting to collect the result set
          schema:
            type: integer
        - in: query
          name: limit
          description: The numbers of items to return
          schema:
            type: integer
      responses:
        "200":
          description: Curated list of movies/series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaListItem"
  /media/genres:
    get:
      summary: List Genres
      description: Returns all known genres
      operationId: listGenres
      tags:
        - Media
      security:
        - permissionAuth: [media:access]
      responses:
        "200":
          description: List of genres, with their label (human readable) and ID (used for the input to any endpoint which allows filtering on genres)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaGenre"
  /media/movie/{id}:
    get:
      summary: Get Movie
      description: Returns the fully inflated DTO for this movie
      operationId: getMovie
      tags:
        - Media
      security:
        - permissionAuth: [media:access]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Movie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
    delete:
      summary: Deletes Movie
      description: Deletes the movie and all it's related transcodes. Any on-going transcodes will be cancelled first.
      operationId: deleteMovie
      tags:
        - Media
      security:
        - permissionAuth: [media:access, media:delete]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Succesfully queued deletion of movie and related transcodes
  /media/series/{id}:
    get:
      summary: Get Series
      description: Returns the fully inflated DTO for this series (which contains the seasons, which further contain their episode stubs)
      operationId: getSeries
      tags:
        - Media
      security:
        - permissionAuth: [media:access]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Series containing all seasons (and their episode stubs)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
    delete:
      summary: Deletes Series
      description: Deletes the series and ALL it's seasons and episodes, including any related transcodes. Any on-going transcodes for the episodes contianed within will be cancelled first.
      operationId: deleteSeries
      tags:
        - Media
      security:
        - permissionAuth: [media:access, media:delete]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Succesfully queued deletion of series/seasons/episodes and related transcodes
  /media/season/{id}:
    delete:
      summary: Deletes Season
      description: Deletes the season and ALL it's episodes, including any related transcodes. Any on-going transcodes for the episodes contianed within will be cancelled first.
      operationId: deleteSeason
      tags:
        - Media
      security:
        - permissionAuth: [media:access, media:delete]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Succesfully queued deletion of season, episodes, and related transcodes
  /media/episode/{id}:
    get:
      summary: Get Episode
      description: Returns the fully inflated DTO for this episode
      operationId: getEpisode
      tags:
        - Media
      security:
        - permissionAuth: [media:access]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Episode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
    delete:
      summary: Deletes Episode
      description: Deletes the episode specified, any related transcodes, and cancels any ongoing transcodes for this episode
      operationId: deleteEpisode
      tags:
        - Media
      security:
        - permissionAuth: [media:access, media:delete]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Successfully queued deletion of episode and related transcodes
  /ingests:
    get:
      summary: List Ingests
      description: Returns all active/troubled ingests
      operationId: listIngests
      tags:
        - Ingests
      security:
        - permissionAuth: [ingest:access]
      responses:
        "200":
          description: List of ingests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingest"
  /ingests/{id}:
    get:
      summary: Get
      description: Returns the ingest with the ID provided
      operationId: getIngest
      tags:
        - Ingests
      security:
        - permissionAuth: [ingest:access]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: The ingest, if found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingest"
    delete:
      summary: Delete
      description: Deletes the ingest task with the ID provided
      operationId: deleteIngest
      tags:
        - Ingests
      security:
        - permissionAuth: [ingest:access, ingest:delete]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Delete successful
  /ingests/{id}/trouble-resolution:
    post:
      summary: Resolve Trouble
      description: Resolves the trouble with the ID provided
      operationId: resolveIngest
      tags:
        - Ingests
      security:
        - permissionAuth: [ingest:access, ingest:modify]
      parameters:
        - $ref: "#/components/parameters/ID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveIngestTroubleRequest"
      responses:
        "200":
          description: Resolution successful
  /ingests/poll:
    post:
      summary: Poll
      description: Requests Ingest Service to re-poll the file system for new files which need to be ingested
      operationId: pollIngests
      tags:
        - Ingests
      security:
        - permissionAuth: [ingest:access, ingest:poll]
      responses:
        "200":
          description: Acknowledged
...
servers:
  - url: http://localhost:8080/api/thea/v1
components:
  securitySchemes:
    permissionAuth:
      type: apiKey
      in: cookie
      name: auth-token
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    IngestTroubleType:
      type: string
      enum: [METADATA_FAILURE, TMDB_FAILURE_UNKNOWN, TMDB_FAILURE_MULTI_RESULT, TMDB_FAILURE_NO_RESULT, UNKNOWN_FAILURE]
    IngestTroubleResolutionType:
      type: string
      enum: [ABORT, RETRY, SPECIFY_TMDB_ID]

    # Ingest Controller DTOs
    IngestTrouble:
      type: object
      required:
        - type
        - message
        - context
        - allowed_resolution_types
      properties:
        type:
          $ref: "#/components/schemas/IngestTroubleType"
        message:
          type: string
        context:
          type: object
        allowed_resolution_types:
          type: array
          items:
            $ref: "#/components/schemas/IngestTroubleResolutionType"
    ResolveIngestTroubleRequest:
      type: object
      required:
        - method
        - context
      properties:
        method:
          $ref: "#/components/schemas/IngestTroubleResolutionType"
        context:
          type: object
          additionalProperties:
            type: string
    Ingest:
      type: object
      required:
        - id
        - path
        - state
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
        state:
          type: string
          enum: [COMPLETE, IDLE, IMPORT_HOLD, INGESTING, TROUBLED]
        trouble:
            $ref: '#/components/schemas/IngestTrouble'
        metadata:
          $ref: '#/components/schemas/FileMetadata'
    FileMetadata:
      type: object
      required:
        - title
        - episodic
        - season_number
        - episode_number
        - runtime
        - path
      properties:
        title:
          type: string
        episodic:
          type: boolean
        season_number:
          type: integer
        episode_number:
          type: integer
        runtime:
          type: string
        year:
          type: integer
        frame_width:
          type: integer
        frame_height:
          type: integer
        path:
          type: string
    MediaWatchTargetType:
      type: string
      enum: ['PRE_TRANSCODE', 'LIVE_TRANSCODE']
    MediaWatchTarget:
      type: object
      required:
        - display_name
        - enabled
        - type
        - ready
      properties:
        display_name:
          type: string
        target_id:
          type: string
          format: uuid
        enabled:
          type: boolean
        type:
          $ref: "#/components/schemas/MediaWatchTargetType"
        ready:
          type: boolean
    Series:
      type: object
      required:
        - id
        - tmdb_id
        - title
        - seasons
      properties:
        id:
          type: string
          format: uuid
        tmdb_id:
          type: string
        title:
          type: string
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/Season"
    Season:
      type: object
      required:
        - episodes
      properties:
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/EpisodeStub"
    Movie:
      type:
        object
      required:
        - id
        - tmdb_id
        - title
        - created_at
        - updated_at
        - watch_targets
      properties:
        id:
          type: string
          format: uuid
        tmdb_id:
          type: string
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        watch_targets:
          type: array
          items:
            $ref: "#/components/schemas/MediaWatchTarget"
    Episode:
      type:
        object
      required:
        - id
        - tmdb_id
        - title
        - created_at
        - updated_at
        - watch_targets
      properties:
        id:
          type: string
          format: uuid
        tmdb_id:
          type: string
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        watch_targets:
          type: array
          items:
            $ref: "#/components/schemas/MediaWatchTarget"
    EpisodeStub:
      type: object
      required:
        - id
        - title
        - adult
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        adult:
          type: boolean
    MediaGenre:
      type: object
      required:
        - id
        - label
      properties:
        id:
          type: string
        label:
          type: string
    MediaListItem:
      type: object
      required:
        - type
        - id
        - title
        - tmdb_id
        - updated_at
        - genres
      properties:
        type:
          type: string
          enum: ['MOVIE', 'SERIES']
        id:
          type: string
          format: uuid
        title:
          type: string
        tmdb_id:
          type: string
        updated_at:
          type: string
          format: date-time
        season_count:
          type: integer
        genres:
          type: array
          items:
            $ref: "#/components/schemas/MediaGenre"
...